# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ArkAPI | Bitcoin-native Pay-per-call API Platform",
    "version": "2026-03-27",
    "description": "Bitcoin-native pay-per-call API platform for developers and AI agents. Access security, OSINT, AI, and utility APIs anonymously using Signet Lightning and Ark. No API keys or accounts required."
  },
  "servers": [
    {
      "url": "https://arkapi.dev",
      "description": "Public ArkAPI endpoint"
    }
  ],
...
  "paths": {
...
    "/v1/sessions": {
      "post": {
        "tags": [
          "public",
          "session"
        ],
        "summary": "Create a new funded session",
        "description": "Creates a new Signet session and returns a bearer token plus Signet Lightning invoice and Signet Ark funding address.",
        "operationId": "createSession",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              },
              "example": {
                "amount_sats": 500
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                },
                "example": {
                  "token": "ak_exampletoken123",
                  "funding": {
                    "lightning_invoice": "lntbs5u1exampleinvoice",
                    "ark_address": "tark1examplefundingaddress",
                    "payment_hash": "examplepaymenthash"
                  },
                  "amount_sats": 500,
                  "balance_sats": 500,
                  "status": "awaiting_payment",
                  "expires_in": 86400
                }
              }
            }
          }
        }
      }
    },
    "/v1/balance": {
      "get": {
        "tags": [
          "session"
        ],
        "summary": "Check session balance",
        "operationId": "getBalance",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Session balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                },
                "example": {
                  "balance_sats": 475,
                  "status": "active",
                  "token": "ak_exampletoken123"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "error": "Unauthorized",
                  "message": "missing or invalid bearer token"
                }
              }
            }
          },
          "402": {
            "description": "Session exists but has insufficient or inactive balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dns-lookup": {
      "post": {
        "tags": [
          "osint"
        ],
        "summary": "Bitcoin-native DNS lookup API",
        "description": "Cost: 3 sats. Returns full, structured DNS records (A, AAAA, MX, NS, TXT, SOA) for a domain as clean JSON. Optimized for AI agents and OSINT automation.",
        "operationId": "dnsLookup",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-cost-sats": 3,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainRequest"
              },
              "example": {
                "domain": "example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful paid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaidSuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "domain": "example.com",
                    "a": [
                      "93.184.216.34"
                    ],
                    "mx": [
                      "0 ."
                    ],
                    "ns": [
                      "a.iana-servers.net.",
                      "b.iana-servers.net."
                    ]
                  },
                  "cost_sats": 3,
                  "balance_remaining": 497,
                  "response_ms": 12,
                  "endpoint": "/api/dns-lookup"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "error": "Unauthorized",
                  "message": "missing or invalid bearer token"
                }
              }
            }
          }
        }
      }
    },
    "/api/axfr-check": {
      "post": {
        "tags": [
          "security"
        ],
        "summary": "Check DNS zone transfer exposure",
        "description": "Cost: 12 sats. Tests whether authoritative nameservers permit AXFR and returns the exposed zone records only when zone transfer is enabled.",
        "operationId": "axfrCheck",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-cost-sats": 12,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AXFRCheckRequest"
              },
              "example": {
                "domain": "example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful paid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaidSuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "domain": "example.com",
                    "allowed": false,
                    "explanation": "AXFR was not permitted by the checked authoritative nameservers, so no zone records were returned.",
                    "nameservers_checked": [
                      "a.iana-servers.net",
                      "b.iana-servers.net"
                    ],
                    "transfer_record_count": 0
                  },
                  "cost_sats": 12,
                  "balance_remaining": 488,
                  "response_ms": 214,
                  "endpoint": "/api/axfr-check"
                }
              }
            }
          }
        }
      }
    },
    "/api/whois": {
      "post": {
        "tags": [
          "osint"
        ],
        "summary": "Parsed WHOIS data API",
        "description": "Cost: 5 sats. Returns cleanly parsed WHOIS registration data (registrar, dates, status) for a domain. Eliminates manual parsing of raw WHOIS output.",
        "operationId": "whoisLookup",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-cost-sats": 5,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainRequest"
              },
              "example": {
                "domain": "example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful paid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaidSuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "domain": "example.com",
                    "registrar": "IANA",
                    "created_date": "1995-08-14",
                    "expiry_date": "2026-08-13"
                  },
                  "cost_sats": 5,
                  "balance_remaining": 495,
                  "response_ms": 18,
                  "endpoint": "/api/whois"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 401,
                  "error": "Unauthorized",
                  "message": "missing or invalid bearer token"
                }
              }
            }
          }
        }
      }
    },
    "/api/ssl-check": {
      "post": {
        "tags": [
          "security"
        ],
        "summary": "SSL certificate analysis",
        "description": "Cost: 5 sats. Returns certificate validity, issuer, protocol, and remaining lifetime.",
        "operationId": "sslCheck",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-cost-sats": 5,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainRequest"
              },
              "example": {
                "domain": "cloudflare.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful paid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaidSuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "valid": true,
                    "issuer": "WE1 / Google Trust Services",
                    "protocol": "TLS 1.3",
                    "days_remaining": 79
                  },
                  "cost_sats": 5,
                  "balance_remaining": 495,
                  "response_ms": 15,
                  "endpoint": "/api/ssl-check"
                }
              }
            }
          }
        }
      }
    },
    "/api/headers": {
      "post": {
        "tags": [
          "security"
        ],
        "summary": "HTTP security headers audit",
        "description": "Cost: 3 sats. Audits a public HTTP or HTTPS URL for security headers. Private, local, and reserved targets are rejected.",
        "operationId": "headersAudit",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-cost-sats": 3,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/URLRequest"
              },
              "example": {
                "url": "https://example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful paid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaidSuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "url": "https://example.com",
                    "grade": "B",
                    "score": 72,
                    "headers": {
                      "strict-transport-security": "max-age=63072000"
                    }
                  },
                  "cost_sats": 3,
                  "balance_remaining": 497,
                  "response_ms": 24,
                  "endpoint": "/api/headers"
                }
              }
            }
          }
        }
      }
    },
    "/api/weather": {
      "post": {
        "tags": [
          "utility"
        ],
        "summary": "Weather forecast",
        "description": "Cost: 3 sats. Returns current conditions and forecast by city or coordinates.",
        "operationId": "weatherLookup",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-cost-sats": 3,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherRequest"
              },
              "example": {
                "city": "New York"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful paid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaidSuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "location": "New York, US",
                    "current": {
                      "temperature_c": 7.2,
                      "condition": "Cloudy"
                    }
                  },
                  "cost_sats": 3,
                  "balance_remaining": 497,
                  "response_ms": 220,
                  "endpoint": "/api/weather"
                }
              }
            }
          }
        }
      }
    },
    "/api/ip-lookup": {
      "post": {
        "tags": [
          "osint"
        ],
        "summary": "IP geolocation and ASN lookup",
        "description": "Cost: 3 sats. Returns country, region, city, coordinates, approximate location, Google Maps link, ISP, org, and AS data for a public IP.",
        "operationId": "ipLookup",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-cost-sats": 3,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IPRequest"
              },
              "example": {
                "ip": "8.8.8.8"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful paid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaidSuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "ip": "8.8.8.8",
                    "country": "United States",
                    "country_code": "US",
                    "region": "Virginia",
                    "city": "Ashburn",
                    "approximate_location": "Ashburn, Virginia, United States",
                    "lat": 39.0438,
                    "lon": -77.4874,
                    "google_maps_url": "https://www.google.com/maps?q=39.0438,-77.4874",
                    "isp": "Google LLC",
                    "org": "Google Public DNS",
                    "as": "AS15169 Google LLC"
                  },
                  "cost_sats": 3,
                  "balance_remaining": 497,
                  "response_ms": 8,
                  "endpoint": "/api/ip-lookup"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/qr-generate": {
      "post": {
        "tags": [
          "utility"
        ],
        "summary": "Generate a QR code PNG",
        "description": "Cost: 2 sats. Returns base64 and data URI output for a PNG QR code.",
        "operationId": "qrGenerate",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-cost-sats": 2,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QRRequest"
              },
              "example": {
                "data": "https://arkapi.dev/fund/"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful paid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaidSuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "format": "png",
                    "size": 256,
                    "data_uri": "data:image/png;base64,iVBORw0KGgoAAA..."
                  },
                  "cost_sats": 2,
                  "balance_remaining": 498,
                  "response_ms": 6,
                  "endpoint": "/api/qr-generate"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ArkAPI session token"
      }
    },
...
    "schemas": {
...
      "CreateSessionRequest": {
        "type": "object",
        "properties": {
          "amount_sats": {
            "type": "integer",
            "example": 500,
            "description": "Requested Signet funding amount"
          }
        }
      },
      "CreateSessionResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "ak_xxx"
          },
          "funding": {
            "type": "object",
            "properties": {
              "lightning_invoice": {
                "type": "string"
              },
              "ark_address": {
                "type": "string"
              },
              "payment_hash": {
                "type": "string"
              }
            }
          },
          "amount_sats": {
            "type": "integer"
          },
          "balance_sats": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "example": "awaiting_payment"
          },
          "expires_in": {
            "type": "integer"
          }
        }
      },
...
      "AXFRCheckRequest": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string",
            "example": "example.com"
          }
        },
        "required": [
          "domain"
        ]
      },
...
    }
  }
}